Add Gemfile (and RSpec, but that's not the point)
Providing a `Gemfile` is a simple yet eficient way to:

- keep track of the app dependencies
- lock their versions!
  (Who knows if our tests will run with RSpec 5.x.x in 8 months?)

IMHO: Any modification of the `Gemfile` MUST be committed with
the resulting `Gemfile.lock`.
Running `bundle install` is enough to generate an up-to-date `Gemfile.lock`.

One exception to that rule: `bundle upgrade <package>` which updates the gems
versions without modification of the `Gemfile`.