Add README
This is the last step, but should be the first. It's one of the most
important files when aiming at peer collaboration. Writing a README is
an unique occasion to show you fellow developers that your care ;)

Ready for some README-driven development?
See http://tom.preston-werner.com/2010/08/23/readme-driven-development.html

Fixes #1